% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dspline.R
\name{expandModelData}
\alias{expandModelData}
\title{Add rows to a model data frame}
\usage{
expandModelData(model, add = NULL)
}
\arguments{
\item{model}{whose model frame is used to obtain
variable names and, particularly, correct
levels for factors used as predictors in the model.}

\item{add}{a data frame with values of predictor variables. The
default, NULL, returns only the model data frame.}
}
\value{
a data frame concatenating the rows of the 
       model frame with those provided in 'add' with
       an additional variable '.source' with value
       'model' or 'add' to indicate the source of the row.
       To get only the added rows, use \code{subset(getD(model,add), .source == 'add')}.
}
\description{
Facilitates creating rows of predictor values to graph
fitted values and standard errors for hypothetical
set of predictor values.
}
