% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspline-package.R
\name{gsplineHexsticker}
\alias{gsplineHexsticker}
\title{View the Official Hex Sticker for the gspline Package}
\usage{
gsplineHexsticker()
}
\description{
Open the official hex sticker for the \pkg{gspline} package in your browser.
}
\examples{
\dontrun{
gsplineHexsticker()
}
}
\author{
Georges Monette and John Fox
}
