% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspline-package.R
\docType{data}
\name{Unemployment}
\alias{Unemployment}
\title{Canadian Monthly Unemployment Rate}
\format{A data frame with 290 observations on the following 2 variables.
\describe{ \item{date}{a \code{"Date"} object.}

\item{unemployment}{Percentage unemployed.} }}
\source{
Downloaded from \url{https://research.stlouisfed.org/} on
2019-03-18.
}
\description{
Monthly unemployment data for Canada for the period 1995-01-01 through
2019-02-01
}
\examples{

 plot(unemployment ~ date, type="l", lwd=2, data=Unemployment)


}
\keyword{datasets}
